Skip local staging in Nexus plugin.
Local staging can have side effects that are sensitive to the sequence of builds (either the sequence of module builds or the sequence of top-level builds), when a multi-module build is restarted from a resumption point. We don't need local staging, and remote staging has well-defined sequencing, so it's simply better to turn local staging off.